requirements/common.txt
decorator==4.0.10
Django==1.10.4
django-bootstrap3==7.1.0
django-filter==1.0.1
django-pjax==1.4
django-registration==2.0.3
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
wcwidth==0.1.7
